(refactor) Modify utils.js name to be more descriptive and update where imported
